requirements.txt
pypdf>=4.0.0
fpdf2>=2.7.0
requests>=2.31.0
pyyaml>=6.0.0
